+++ src/main/scala/org/nd4s/OperatableNDArray.scala
@@ -17,7 +17,7 @@
  *
  */
 
-package org.nd4j.api
+package org.nd4s
 
 import org.nd4j.linalg.api.complex.{IComplexNDArray, IComplexNumber}
 import org.nd4j.linalg.api.ndarray.INDArray
