+++ src/main/scala/org/nd4s/SliceableNDArray.scala
@@ -1,6 +1,6 @@
-package org.nd4j.api
+package org.nd4s
 
-import org.nd4j.api.Implicits._
+import org.nd4s.Implicits._
 import org.nd4j.linalg.api.complex.{IComplexNDArray, IComplexNumber}
 import org.nd4j.linalg.api.ndarray.INDArray
 import org.nd4j.linalg.factory.Nd4j
