+++ src/test/scala/org/nd4s/NDArrayExtractionTest.scala
@@ -1,6 +1,6 @@
-package org.nd4j.api
+package org.nd4s
 
-import org.nd4j.api.Implicits._
+import org.nd4s.Implicits._
 import org.nd4j.linalg.api.complex.IComplexNDArray
 import org.nd4j.linalg.factory.Nd4j
 import org.scalatest.FlatSpec
